timers/schedulers in core
