+++ src/Proto.Cluster/Partition/PartitionMemberSelector.cs
@@ -12,10 +12,7 @@ namespace Proto.Cluster.Partition
     internal class PartitionMemberSelector
     {
         private readonly object _lock = new();
-        private readonly List<Member> _members = new();
         private readonly Rendezvous _rdv = new();
-
-        public int Count => _members.Count;
 
         public void Update(Member[] members)
         {
